Updated dialogue box and health settings

+++ FinalProject/src/Actor.java
@@ -2,26 +2,26 @@ import javafx.scene.canvas.Canvas;
 
 public class Actor extends Canvas {
 //	private Image sprite;
-	private int xPos;
-	private int yPos;
+	private double xPos;
+	private double yPos;
 	
 	public Actor(int width, int height) {
 		super(width, height);
 	}
 
-	public int getxPos() {
+	public double getxPos() {
 		return xPos;
 	}
 
-	public void setxPos(int xPos) {
+	public void setxPos(double xPos) {
 		this.xPos = xPos;
 	}
 
-	public int getyPos() {
+	public double getyPos() {
 		return yPos;
 	}
 
-	public void setyPos(int yPos) {
+	public void setyPos(double yPos) {
 		this.yPos = yPos;
 	}
